So here's what happens:
Every 5 second, fetchprojectposts.php will be called and the screen will be updated. It will also be called when the user posts a new message.

Possible solution to try-> ping a page that will display update when anything new happens. If there is anything to be retrived, fetchprojectposts.php then actually runs. TAHDAH. not hard to implement but I want my sleep. Good night